+++ sdmx-dl-provider-ri/src/test/java/internal/sdmxdl/provider/ri/web/drivers/BbkDriverTest.java
@@ -59,11 +59,11 @@ public class BbkDriverTest {
                 .describedAs("SdmxFix#1 + SdmxFix#2")
                 .hasToString("https://api.statistiken.bundesbank.de/rest/metadata/datastructure/BBK/BBK_ERX?references=children");
 
-        assertThat(queries.getDataQuery(endpoint, DataRef.of(DataflowRef.parse("BBEX3"), DataQuery.of(Key.parse("M.ISK.EUR+USD.CA.AC.A01"), FULL))).build())
+        assertThat(queries.getDataQuery(endpoint, DataRef.of(DataflowRef.parse("BBEX3"), DataQuery.of(Key.parse("M.ISK.EUR+USD.CA.AC.A01"), FULL)), DataStructureRef.parse("abc")).build())
                 .describedAs("SdmxFix#4")
                 .hasToString("https://api.statistiken.bundesbank.de/rest/data/BBEX3/M.ISK.EUR%2BUSD.CA.AC.A01");
 
-        assertThat(queries.getDataQuery(endpoint, DataRef.of(DataflowRef.parse("BBEX3"), DataQuery.of(Key.parse("M.ISK.EUR+USD.CA.AC.A01"), SERIES_KEYS_ONLY))).build())
+        assertThat(queries.getDataQuery(endpoint, DataRef.of(DataflowRef.parse("BBEX3"), DataQuery.of(Key.parse("M.ISK.EUR+USD.CA.AC.A01"), SERIES_KEYS_ONLY)), DataStructureRef.parse("abc")).build())
                 .describedAs("SdmxFix#5")
                 .hasToString("https://api.statistiken.bundesbank.de/rest/data/BBEX3/M.ISK.EUR%2BUSD.CA.AC.A01?detail=serieskeyonly");
     }